+++ internal/admission/handler.go
@@ -23,6 +23,7 @@ import (
 const (
 	KindKongPlugin        = "KongPlugin"
 	KindKongClusterPlugin = "KongClusterPlugin"
+	KindGateway           = gatewayapi.Kind("Gateway")
 )
 
 // RequestHandler is an HTTP server that can validate Kong Ingress Controllers'